+++ CHANGELOG.md
@@ -15,6 +15,7 @@ No public interface changes since `22.3.0`.
 
 - Fixed bug in `EuiAccordion` to adjust to the correct height when content height changes ([#3160](https://github.com/elastic/eui/pull/3160))
 - Fixed bug in `EuiBasicTable` to handle dynamic icon value properly in collapsed actions ([#3145](https://github.com/elastic/eui/pull/3145))
+- Fixed `availability` check for actions in `EuiBasicTable` ([3030](https://github.com/elastic/kibana/issues/3030))
 
 ## [`22.2.0`](https://github.com/elastic/eui/tree/v22.2.0)